Added very early custom compression (WIP)
This uses basic Huffman coding on the screen data, and DEFLATE on colors.
This alone manages to save about 50% of the data size (which is comparable
to DEFLATE over the entire file).
This is not ready for production - DEFLATE is *still* too slow. Might
RLE or Huffman code the colors instead.