Added GPU k-means; other fixes
* Fixed k-means not working properly on CPU
* Fixed various out-of-bounds errors in GPU code
* Fixed crustiness in GPU CC pixel output

WARNING: Floyd-Steinberg dithering appears to be HORRIBLY BROKEN on GPU
for some reason - it hangs my entire system for a minute. Basically the
same code worked before, and I'm not sure why it's not anymore, but I'll
have to do a lot more debugging to get it to not freeze everything.